+++ src/Text/LLVM/PP.hs
@@ -89,6 +89,8 @@ module Text.LLVM.PP
   , ppCallSym
   , ppGEP
   , ppInvoke
+  , ppOperandBundles
+  , ppOperandBundle
   , ppPhiArg
   , ppICmpOp
   , ppFCmpOp
@@ -810,8 +812,8 @@ ppInstr instr = case instr of
                          <> comma <+> ppValue r
   Conv op a ty           -> ppConvOp op <+> ppTyped ppValue a
                         <+> "to" <+> ppType ty
-  Call tc ty f args      -> ppCall tc ty f args
-  CallBr ty f args u es  -> ppCallBr ty f args u es
+  Call tc ty f args bs   -> ppCall tc ty f args bs
+  CallBr ty f args u es bs -> ppCallBr ty f args u es bs
   Alloca ty len align    -> ppAlloca ty len align
   Load ty ptr mo ma      -> ppLoad ty ptr mo ma
   Store a ptr mo ma      -> ppStore a ptr mo ma
@@ -857,7 +859,7 @@ ppInstr instr = case instr of
                         <+> ppLabel t
                          <> comma <+> ppType (PrimType Label)
                         <+> ppLabel f
-  Invoke ty f args to uw -> ppInvoke ty f args to uw
+  Invoke ty f args to uw bs -> ppInvoke ty f args to uw bs
   Unreachable            -> "unreachable"
   Unwind                 -> "unwind"
   VaArg al t             -> "va_arg" <+> ppTyped ppValue al
@@ -992,20 +994,22 @@ ppAlloca ty mbLen mbAlign = "alloca" <+> ppType ty <> len <> align
     a <- mbAlign
     return (comma <+> "align" <+> int a)
 
-ppCall :: Bool -> Type -> Value -> Fmt [Typed Value]
-ppCall tc ty f args
+ppCall :: Bool -> Type -> Value -> [Typed Value] -> Fmt [OperandBundle]
+ppCall tc ty f args bs
   | tc        = "tail" <+> body
   | otherwise = body
   where
   body = "call" <+> ppCallSym ty f
       <> parens (commas (map (ppTyped ppValue) args))
+     <+> ppOperandBundles bs
 
 -- | Note that the textual syntax changed in LLVM 10 (@callbr@ was introduced in
 -- LLVM 9).
-ppCallBr :: Type -> Value -> [Typed Value] -> BlockLabel -> Fmt [BlockLabel]
-ppCallBr ty f args to indirectDests =
+ppCallBr :: Type -> Value -> [Typed Value] -> BlockLabel -> [BlockLabel] -> Fmt [OperandBundle]
+ppCallBr ty f args to indirectDests bs =
   "callbr"
      <+> ppCallSym ty f <> parens (commas (map (ppTyped ppValue) args))
+     <+> ppOperandBundles bs
      <+> "to" <+> ppLab to <+> brackets (commas (map ppLab indirectDests))
   where
     ppLab l = ppType (PrimType Label) <+> ppLabel l
@@ -1049,13 +1053,28 @@ ppGEP gf ty ptr ixs =
 
   explicit = ppType ty <> comma
 
-ppInvoke :: Type -> Value -> [Typed Value] -> BlockLabel -> Fmt BlockLabel
-ppInvoke ty f args to uw = body
+ppInvoke :: Type -> Value -> [Typed Value] -> BlockLabel -> BlockLabel -> Fmt [OperandBundle]
+ppInvoke ty f args to uw bs = body
   where
   body = "invoke" <+> ppCallSym ty f
       <> parens (commas (map (ppTyped ppValue) args))
+     <+> ppOperandBundles bs
      <+> "to" <+> ppType (PrimType Label) <+> ppLabel to
      <+> "unwind" <+> ppType (PrimType Label) <+> ppLabel uw
+
+-- | Pretty-print a list of operand bundles as @[ "tag"(args), "tag2"(args) ]@,
+-- or 'empty' if the list is empty. Each bundle's args list may itself be empty
+-- (e.g. @[ "deopt"() ]@). The verifier requires bundle tags to be non-empty,
+-- but we don't enforce that here.
+ppOperandBundles :: Fmt [OperandBundle]
+ppOperandBundles []  = empty
+ppOperandBundles bs  =
+  brackets (commas (map ppOperandBundle bs))
+
+ppOperandBundle :: Fmt OperandBundle
+ppOperandBundle (OperandBundle tag args) =
+  doubleQuotes (text tag)
+    <> parens (commas (map (ppTyped ppValue) args))
 
 ppPhiArg :: Fmt (Value,BlockLabel)
 ppPhiArg (v,l) = char '[' <+> ppValue v <> comma <+> ppLabel l <+> char ']'